college-planner/agent/pyproject.toml
[tool.poetry]
name = "college_planner"
version = "0.1.0"
description = "A college savings planner"
authors = ["Kia Afzali"]
readme = "README.md"
exclude = [
    "college_planner/suql/faiss_embedding.py",
    # Add other patterns here if needed
]


[tool.poetry.dependencies]
python = "^3.12"
langchain-openai = "^0.2.1"
langchain-anthropic = "^0.2.1"
langchain = "^0.3.1"
openai = "^1.51.0"
langchain-community = "^0.3.1"
copilotkit = "0.1.34"
uvicorn = "^0.31.0"
python-dotenv = "^1.0.1"
#tavily-python = "^0.5.0"
html2text = "^2024.2.26"
googlemaps = "^4.10.0"
langgraph-cli = {extras = ["inmem"], version = "^0.1.64"}
langchain-core = "^0.3.25"


pglast = "5.3.0"
sympy = "^1.13.1"
sqlparse = "0.5.3" 

Jinja2 = "3.1.2"
Flask = "2.3.2"
Flask-Cors = "4.0.0"
Flask-RESTful = "0.3.10"
requests = "2.31.0"
tiktoken = "0.7.0"
psycopg2-binary = "2.9.9"
litellm = "1.34.34"
platformdirs = "4.0.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
demo = "college_planner.demo:main"
